refactored MakeLocalVariableConstantTest.cs
